You can now modify the amount of items in the inventory
You must first have that item in the inventory to decrement the number
of items you have. If you want to decrement, you must type in a negative
number. You can also add items, there is not currently a check to see if
there are enough items currently in the inventory to remove them